[modify]extract the proxy of Itestmethod
